move action into sites controller
