+++ config/routes.rb
@@ -6,7 +6,7 @@ Alchemy::Engine.routes.draw do
   get '/sitemap.xml' => 'pages#sitemap', format: 'xml'
 
   get '/admin' => redirect('admin/dashboard')
-  get '/admin/sites/:id/select' => 'admin/change_site#index',
+  get '/admin/sites/:id/select' => 'admin/sites#select',
         :as => :admin_select_site
   get '/admin/dashboard' => 'admin/dashboard#index',
         :as => :admin_dashboard
