+++ spec/controllers/admin/sites_controller_spec.rb
@@ -2,12 +2,12 @@ require 'ostruct'
 require 'spec_helper'
 
 module Alchemy
-  describe Admin::ChangeSiteController do
+  describe Admin::SitesController do
     let!(:a_site) { FactoryGirl.create(:site) }
 
     context 'a guest' do
       it 'can not access site switch' do
-        alchemy_get :index, id: a_site.id
+        alchemy_get :select, id: a_site.id
         expect(response).to be_redirect
       end
     end
@@ -16,7 +16,7 @@ module Alchemy
       before { authorize_user(build(:alchemy_dummy_user)) }
 
       it 'can not access site switch' do
-        alchemy_get :index, id: a_site.id
+        alchemy_get :select, id: a_site.id
         expect(response).to be_redirect
       end
     end
@@ -27,7 +27,7 @@ module Alchemy
 
       it 'can access site switch' do
         expect{
-          alchemy_get :index, redirect_to: "http://www.google.de", id: a_site.id
+          alchemy_get :select, redirect_to: "http://www.google.de", id: a_site.id
         }.to change{ session[:site_id] }.to a_site.id.to_s
         expect(response).to redirect_to("http://www.google.de")
       end
